include: add alignment helper header
Add some helper function to provide alignment functionality.

ALIGN will align up a base "a" with respect to "b".
For example, a = 0x0B, b = 0x4, ALIGN(a, b) = 0x0C

ALIGN_DOWN will align down a base "a" with respect to "b".
For example, a = 0x0B, b = 0x4, ALIGN_DOWN(a, b) = 0x08

Note that, "b" must be a power of 2.

Signed-off-by: daitor <tim90721@gmail.com>