risc-v: implement trap handler for external interrupt
Enable external interrupt trap handler.
To enable this function, we must enable global level interrupt enable in
MSTATUS and enable external interrupt enable in MIE.
Whenever a trap occurred, we can identify this trap is an interrupt or
exception via MCAUSE's MSB. If the MSB is 1, it means that the trap is an
interrupt and we can hand this trap to architecture level interrupt
handler.

Signed-off-by: daitor <tim90721@gmail.com>