risc-v: add support to handle timer IRQ
Add support to handle timer IRQ.
CLINT has per CPU counter which can be used to generated regular timer IRQ
to maintain a per-CPU timer.

The timer IRQ is generated from CLINT and route to per-CPU trap. For
RISC-V, MTIE needs to be enabled to receive the timer IRQ. Moreover, The
MCAUSE will carry a special IRQ ID to represent the IRQ is belonged to
timer.

The regular timer is generated by feeding the CLINT per-CPU counter with
additional HZ count. When CLINT MTIME exceeds MTIMECMP, it will generate
the timer IRQ. By feeding the MTIMECMP with a value greater than MTIME,
the IRQ can be considered to IRQ handled done.

Signed-off-by: daitor <tim90721@gmail.com>