arm64: implement interrupt handler
Implement bare metal interrupt handler.
The handler will save the context on top of the current task's context
which may occupy some stack. There is a potential stack overflow need to
be cared about and it is needed to be fixed in the future.

Armv8 interrupt and exception table base address is placed to VBAR_EL1 and
the base address must be aligned to 2048B address. Moreover, the table is
divided to 4 parts which dedicates to different EL and SP use case. Each
part contains 4 blocks which needs to be aligned to 128B. In other words,
each block is able to accommodate 32 instructions.

For now, this commit only implement the second part of the table and it
only supports interrupt handling.

Signed-off-by: tim90721 <tim90721@gmail.com>